+++ src/histoseg/contours/__init__.py
@@ -1,2 +1,0 @@
-from .multi_structure import *
-from .pattern1_isoline import *